+++ provider/mysql/mysql.go
@@ -12,14 +12,14 @@ type mysqlProvider struct {
 	connectionString string
 	timeout          time.Duration
 	interval         time.Duration
-	downThreshold    int
-	upThreshold      int
+	downThreshold    int64
+	upThreshold      int64
 }
 
 func NewMysqlProvider(
 	name, connectionString string,
 	timeout, interval time.Duration,
-	downThreshold, upThreshold int,
+	downThreshold, upThreshold int64,
 ) provider.Provider {
 	return &mysqlProvider{
 		name:             name,
@@ -39,11 +39,11 @@ func (mp mysqlProvider) GetInterval() time.Duration {
 	return mp.interval
 }
 
-func (mp mysqlProvider) GetDownThreshold() int {
+func (mp mysqlProvider) GetDownThreshold() int64 {
 	return mp.downThreshold
 }
 
-func (mp mysqlProvider) GetUpThreshold() int {
+func (mp mysqlProvider) GetUpThreshold() int64 {
 	return mp.upThreshold
 }
